chore: use Revision

+++ packages/rs-dpp/src/document/state_transition/documents_batch_transition/validation/state/validate_documents_batch_transition_state.rs
@@ -250,7 +250,7 @@ fn check_revision(
         result.add_error(ConsensusError::StateError(Box::new(
             StateError::InvalidDocumentRevisionError {
                 document_id: document_transition.base().id.clone(),
-                current_revision: fetched_document.revision,
+                current_revision: fetched_document.revision as u64,
             },
         )))
     }
